Create Dynamic Data Structure & Linked List Question 3
Content :
Based on the program main() written in 2, write a function searchList() that examines a linked list of integers to determine if a given integer value is stored within the list. searchList() is called in main() once the linked list is built and printed.

         int searchList(ListNode *head, int value);

The function should return the index of the first appearance node containing the value if it is found and -1 otherwise. Sample output:

Current list: 1 2 3 5 6

Enter value to search for: 5

Value 5 found at index 3