Create Filter Odd Digits from Queue
Write a C function createQueueFromLinkedList() to create a queue (linked-list-based) by
enqueuing all integers which are stored in the linked list. The first node of the linked list is
enqueued first, and then the second node, and so on. Remember to empty the queue at the
beginning, if the queue is not empty.
After the queue is built, write another C function removeOddValues() to remove all odd
integers from the queue. Note that you should only use enqueue() or dequeue() when you
add or remove integers from queues.
The function prototypes are given as follows:
void createQueueFromLinkedList(LinkedList *ll , Queue *q);
void removeOddValues(Queue *q)
A sample input and output session is given below (if the current linked list is 1 2 3 4 5):
The resulting linked list is: 1 2 3 4 5
Please input your choice(1/2/3/0): 2
The resulting queue is: 1 2 3 4 5
Please input your choice(1/2/3/0): 3
The resulting queue after removing odd integers is: 2 4